#27 Refactor: 팀 가입 신청 기능 리팩토링
