+++ wtnt/team/apply/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views import TeamApplyView
+
+urlpatterns = [
+    path("apply/<int:team_id>", TeamApplyView.as_view()),
+]
